tests/infra/db/mongodb/account-mongo-repository.spec.ts
import { Collection } from 'mongodb'
import faker from '@faker-js/faker'

import { MongoHelper, AccountMongoRepository } from '@/infra/db'
import { mockAddAccountParams } from '@/tests/domain/mocks'

let accountCollection: Collection

const makeSut = (): AccountMongoRepository => {
  return new AccountMongoRepository()
}

describe('AccountMongoRepository', () => {
  beforeAll(async () => {
    await MongoHelper.connect(process.env.MONGO_URL)
  })

  beforeEach(async () => {
    accountCollection = MongoHelper.getCollection('accounts')
    await accountCollection.deleteMany({})
  })

  afterAll(async () => {
    await MongoHelper.disconnect()
  })

  describe('add()', () => {
    it('should add account on success', async () => {
      const sut = makeSut()
      const account = await sut.add(mockAddAccountParams())
      expect(account).toBe(true)
    })
  })

  describe('checkByEmail()', () => {
    it('should return true if email is valid', async () => {
      const sut = makeSut()
      const data = mockAddAccountParams()
      await accountCollection.insertOne(data)
      const account = await sut.checkByEmail(data.email)
      expect(account).toBe(true)
    })

    it('should return false if email is not valid', async () => {
      const sut = makeSut()
      await accountCollection.insertOne(mockAddAccountParams())
      const account = await sut.checkByEmail(faker.internet.email())
      expect(account).toBe(false)
    })
  })
})
